Close the branch mutation gate's twenty-two survivors
The branch gate scored 98.2% over 1290 mutants; the 22 survivors sat
in four files this branch touched, and each is now killed or proven
equivalent.

numbered-statement.ts: nothing pinned allConditions or ALWAYS_TRUE.
Three direct cases now hold the empty, single, and joined shapes.

payment-finalize.ts: the free-session finalize now checks the blind
index stays empty with its reference, and a new case pins the balance
finalize's bound attendee-id slot. The two ?? fallbacks are recorded
as equivalent: each left side is a string or undefined with an empty
fallback, so ?? and || agree on every value.

payment-reference-store.ts: new cases pin the blind identity sets for
tagged and untagged references, the empty-index row that loads without
complaint, the mismatched index that refuses loudly with its exact
message, and the unreadable reference that names its column.

groups.ts: its membership-write tests sat in membership.test.ts, which
mirrors groups/membership.ts — so the gate never ran them for
groups.ts. The setListingGroups, setListingGroupsTx, and
setGroupPackageMembers tests move to set-groups.test.ts, which
groups.ts owns; membership.test.ts keeps the assignListingsToGroup
batch tests. The shared group-ids read moves to a test helper.
